feat: move to ESM
Now the package is a native ES module

BREAKING CHANGE: NodeJS >= 14 is required, no commonjs support
