+++ src/browsers/index.ts
@@ -1,5 +1,5 @@
-export * from './types'
-export * from './util'
-export * from './shirtnames'
-export * from './browserslist'
-export * from './optimize'
+export * from './types.js'
+export * from './util.js'
+export * from './shirtnames.js'
+export * from './browserslist.js'
+export * from './optimize.js'
