+++ src/index.ts
@@ -1,5 +1,5 @@
-export * from './useragent'
-export * from './browsers'
-export * from './semver'
-export * from './regexp'
-export * from './useragentRegexp'
+export * from './useragent/index.js'
+export * from './browsers/index.js'
+export * from './semver/index.js'
+export * from './regexp/index.js'
+export * from './useragentRegexp/index.js'
