+++ src/regexp/index.ts
@@ -1,7 +1,7 @@
-export * from './util'
-export * from './optimize'
-export * from './numberRay'
-export * from './numberSegment'
-export * from './numberRange'
-export * from './numbersPart'
-export * from './versions'
+export * from './util.js'
+export * from './optimize.js'
+export * from './numberRay.js'
+export * from './numberSegment.js'
+export * from './numberRange.js'
+export * from './numbersPart.js'
+export * from './versions.js'
