+++ src/useragent/index.ts
@@ -1,4 +1,4 @@
-export * from './types'
-export * from './util'
-export * from './versions'
-export * from './useragent'
+export * from './types.js'
+export * from './util.js'
+export * from './versions.js'
+export * from './useragent.js'
