+++ src/useragentRegexp/index.ts
@@ -1,4 +1,4 @@
-export * from './types'
-export * from './optimize'
-export * from './workarounds'
-export * from './useragentRegexp'
+export * from './types.js'
+export * from './optimize.js'
+export * from './workarounds.js'
+export * from './useragentRegexp.js'
