+++ vite.config.js
@@ -1,0 +1,9 @@
+import { defineConfig } from 'vite'
+
+export default defineConfig({
+  test: {
+    coverage: {
+      reporter: ['lcovonly', 'text']
+    }
+  }
+})
